+++ tests/test_commands_blender_render_progressive.py
@@ -71,4 +71,5 @@ class BlenderRenderProgressiveTest(AbstractCommandTest):
                 stdin=subprocess.DEVNULL,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.STDOUT,
+                env=mock.ANY,
             )